feat(batch): 새 플레이어 업데이트 잡 로직 수정

+++ batch/src/main/java/com/imstargg/batch/job/support/QuerydslPagingItemReader.java
@@ -69,7 +69,7 @@ public class QuerydslPagingItemReader<T> extends AbstractPagingItemReader<T> {
         EntityTransaction tx = getTxOrNull();
 
         JPQLQuery<T> query = createQuery()
-                .offset((long) getPage() * getPageSize())
+                .offset(getPage() * getPageSize())
                 .limit(getPageSize());
 
         initResults();
